MainActivity and DetailsActivity were migrated to fragments added toast about exit
