+++ .idea/misc.xml
@@ -13,6 +13,8 @@
         <entry key="app/src/main/res/layout/activity_details.xml" value="0.33" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.33" />
         <entry key="app/src/main/res/layout/film_item.xml" value="0.33" />
+        <entry key="app/src/main/res/layout/fragment_details.xml" value="0.25" />
+        <entry key="app/src/main/res/layout/fragment_home.xml" value="0.25" />
         <entry key="app/src/main/res/menu/navigation.xml" value="0.24048871015521164" />
         <entry key="app/src/main/res/menu/toolbar.xml" value="0.2296875" />
         <entry key="app/src/main/res/menu/top_app_bar.xml" value="0.2296875" />
