+++ app/src/main/AndroidManifest.xml
@@ -8,9 +8,9 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.KinoProisk">
+        android:theme="@style/AppTheme">
         <activity
-            android:name=".DetailsActivity"
+            android:name=".DetailsFragment"
             android:exported="true" />
         <activity
             android:name=".MainActivity"
