+++ app/src/main/res/layout/fragment_details.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".DetailsActivity"
+    tools:context=".DetailsFragment"
     android:id="@+id/details_ac">
 
     <com.google.android.material.appbar.AppBarLayout
@@ -60,23 +60,24 @@
         android:id="@+id/details_later"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_margin="75dp"
+        android:layout_marginRight="20dp"
         android:backgroundTint="@color/colorPrimaryDark"
         android:tint="@android:color/white"
         app:layout_anchor="@id/app_bar"
-        app:layout_anchorGravity="bottom|right"
+        app:layout_anchorGravity="bottom|center"
         app:srcCompat="@drawable/ic_round_watch_later" />
 
     <com.google.android.material.floatingactionbutton.FloatingActionButton
         android:id="@+id/details_tofav"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_margin="130dp"
+        android:layout_margin="20dp"
         android:backgroundTint="@color/colorPrimaryDark"
+        android:tint="@android:color/white"
         app:layout_anchor="@id/app_bar"
-        app:layout_anchorGravity="bottom|right"
+        app:layout_anchorGravity="bottom|left"
         app:srcCompat="@drawable/ic_round_favorite"
-        android:tint="@android:color/white"/>
+        />
 
     <androidx.core.widget.NestedScrollView
         android:layout_width="match_parent"
