+++ app/src/main/res/values/strings.xml
@@ -13,4 +13,5 @@
     <string name="More">More</string>
     <string name="page_title">Поищем</string>
     <string name="refresh">refresh</string>
+    
 </resources>
